cr frontend

+++ src/app/api/projects/[id]/route.ts
@@ -3,11 +3,11 @@ import { connectToDb } from "@/lib/connectToDb";
 import { IProject, Project, zProject } from "@/models/project.model";
 import { NextRequest, NextResponse } from "next/server";
 
-export const GET = async (
-  req: NextRequest,
-  { params }: any
-): Promise<NextResponse> => {
+export const GET = auth(async (req, { params }: any): Promise<NextResponse> => {
   try {
+    if (!req.auth) {
+      return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+    }
     await connectToDb();
     const { id } = await params;
     const project = await Project.findById(id);
@@ -15,36 +15,40 @@ export const GET = async (
   } catch (error) {
     return NextResponse.json(error, { status: 500 });
   }
-};
+});
 
-export const PATCH = async (
-  req: NextRequest,
-  { params }: any
-): Promise<NextResponse> => {
-  try {
-    await connectToDb();
-    const { id } = await params;
-    const body: IProject = await req.json();
-    zProject.parse(body);
-    const updatedProject = await Project.findByIdAndUpdate(id, body, {
-      new: true,
-    });
-    return NextResponse.json(updatedProject, { status: 201 });
-  } catch (error) {
-    return NextResponse.json(error, { status: 500 });
+export const PATCH = auth(
+  async (req, { params }: any): Promise<NextResponse> => {
+    try {
+      if (!req.auth) {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+      }
+      await connectToDb();
+      const { id } = await params;
+      const body: IProject = await req.json();
+      zProject.parse(body);
+      const updatedProject = await Project.findByIdAndUpdate(id, body, {
+        new: true,
+      });
+      return NextResponse.json(updatedProject, { status: 201 });
+    } catch (error) {
+      return NextResponse.json(error, { status: 500 });
+    }
   }
-};
+);
 
-export const DELETE = async (
-  req: NextRequest,
-  { params }: any
-): Promise<NextResponse> => {
-  try {
-    await connectToDb();
-    const { id } = await params;
-    const deletedProject = await Project.findByIdAndDelete(id, { new: true });
-    return NextResponse.json(deletedProject, { status: 200 });
-  } catch (error) {
-    return NextResponse.json(error, { status: 500 });
+export const DELETE = auth(
+  async (req, { params }: any): Promise<NextResponse> => {
+    try {
+      if (!req.auth) {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+      }
+      await connectToDb();
+      const { id } = await params;
+      const deletedProject = await Project.findByIdAndDelete(id, { new: true });
+      return NextResponse.json(deletedProject, { status: 200 });
+    } catch (error) {
+      return NextResponse.json(error, { status: 500 });
+    }
   }
-};
+);
